+++ lib/duck_puncher/duck.rb
@@ -5,6 +5,8 @@ module DuckPuncher
     # @param target [String,Class] Class or module to punch
     # @param mod [String,Module] The module that defines the extensions
     # @param [Hash] options to modify the duck #punch method behavior
+    # @option options :only [Symbol, Array<Symbol>] list of methods to extend onto the target (the module must have these defined)
+    # @option options :method [Symbol,String] the method used to apply the module, e.g. :extend (:include)
     # @option options :before [Proc] A hook that is called with the target class before +punch+
     # @option options :after [Proc] A hook that is called with the target class after +punch+
     def initialize(target, mod, options = {})
@@ -14,9 +16,9 @@ module DuckPuncher
     end
 
     # @param [Hash] opts to modify punch
-    # @option options [Class] :target Specifies the class to be punched
-    # @option options [Array,Symbol] :only Specifies the methods to extend onto the current object
-    # @option options [Symbol,String] :method Specifies if the methods should be included or prepended (:include)
+    # @option options :target [Class] Specifies the class to be punched (overrides @target)
+    # @option options :only [Array,Symbol] Specifies the methods to extend onto the current object
+    # @option options :method [Symbol,String] Specifies if the methods should be included or prepended (:include)
     # @return [Class] The class that was just punched
     def call(opts = {})
       opts = options.merge(opts)